refactor: add constant for server error
