+++ src/errors/constants.js
@@ -1,5 +1,6 @@
 const ERRORS={
   NOT_FOUND: 404,
+  SERVER_ERROR: 500,
 }
 
 module.exports = ERRORS;
