+++ src/errors/errors.js
@@ -1,5 +1,7 @@
+const Errors = require('./constants');
+
 class ErrorDefiner extends Error {
-  constructor(message = 'Not found', status=500) {
+  constructor(message = 'Not found', status = Errors.NOT_FOUND) {
     super(message);
     this.status = status;
   }
